new(demo): refactor to animated codesandbox pie example
