+++ packages/vx-demo/src/pages/Pies.tsx
@@ -1,21 +1,11 @@
 import React from 'react';
 import Show from '../components/Show';
-import Pies from '../components/tiles/Pies';
-import PiesSource from '!!raw-loader!../components/tiles/Pies';
+import Pies from '../docs-v2/examples/vx-shape-pie/Example';
+import PiesSource from '!!raw-loader!../docs-v2/examples/vx-shape-pie/Example';
 
 export default () => {
   return (
-    <Show
-      events
-      margin={{
-        top: 10,
-        left: 40,
-        right: 30,
-        bottom: 80,
-      }}
-      component={Pies}
-      title="Pies"
-    >
+    <Show component={Pies} title="Pies" codeSandboxDirectoryName="vx-shape-pie">
       {PiesSource}
     </Show>
   );
